Add Info bar

+++ FreeAPS/Sources/Charts/Models/InformationBarEntryData.swift
@@ -3,6 +3,6 @@ import Foundation
 struct InformationBarEntryData: Identifiable, Hashable {
     var id = UUID()
     let label: String
-    let value: Double
+    let value: Decimal
     let type: APSDataTypes
 }
